Fitting Additive partial linear models with symmetric errors

aplms fits additive partial linear models with autoregressive symmetric errors. This method is suitable for data sets where the response variable is continuous and symmetric, with either heavy or light tails, and measured over time. The model includes a parametric component for a set of covariates, while another set of covariates can be specified as semi-parametric functions, typically time-related. In this setup, natural cubic splines or cubic P-splines are used to approximate the nonparametric components.

Usage

aplms(
  formula,
  npc,
  basis,
  Knot,
  data,
  family = Normal(),
  p = 1,
  control = list(tol = 0.001, algorithm1 = c("P-GAM"), algorithm2 = c("BFGS"), Maxiter1 =
    20, Maxiter2 = 25),
  init,
  lam
)

# S3 method for class 'aplms'
print(x, ...)

Arguments

formula

A symbolic description of the parametric component of the model to be fitted. The details of model specification are given under ‘Details’.

npc

A vector of names of non parametric component.

basis

A vector of names of the basis to be used for each non parametric covariate.

Knot

A vector of the number of knots in each non-linear component of the model.

data

A data frame containing the variables in the model

family

Symmetric error distribution. The implemented distribution are: Normal(), LogisI(), LogisII(), Student(df), Powerexp(k), Gstudent(parm=c(s,r)).

p

autoregressive order of the error

control

optimization rutine.

init

A list of initial values for the symmetric error scale, phi, and autoregressive coefficients, rhos.

lam

smoothing parameter vector.

x

APLMS object.

...

other arguments

Value

Returns an object of class “aplms”, a list with following components.

formula

the formula object used.

family

the family object used.

npc

the npc object used.

Knot

the Knot object used.

lam

the lam object used.

rdf

Degrees of freedom: n - q - p - 1.

VAR_F

Estimate the asymptotic covariance matrix for the gamma parameters.

basis

The basis to be used for each non parametric covariate.

WALD_f

The summary table of the Wald statistics.

summary_table_phirho

The summary table of the rho and phi parameters.

N_i

Basis functions.

f

Estimated gamma parameters.

Dv

Dv values for the symmetric error.

Dm

Dm values for the symmetric error.

Dc

Dc values for the symmetric error.

Dd

Dd values for the symmetric error.

delta

delta_i for the symmetric error.

LL_obs

Observed information matrix of the fitted model.

loglike

The estimated loglikelihood function of the fitted model.

total_df

The total effective degree of freedom of the model.

parametric_df

The degree of freedom of the parametric components.

npc_df

The effective degree of freedom of the non parametric components.

AIC

Akaike information criterion of the estimated model.

BIC

Bayesian information criterion of the estimated model.

AICC

Corrected Akaike information criterion of the estimated model.

GCV

The generalized cross-validation (GCV).

yhat

The fitted response values of the model.

muhat

The fitted mean values of the model.

residuals_y

The response residuals

residuals_mu

Raw (Ordinary) residuals: \(y_t - (\textbf{x}_i^\top\beta + f_1(t_{i1}) + \ldots + f_k(t_{ik}))\)

data

the data object used.

this.call

the function call used.

References

Chou-Chen, S.W., Oliveira, R.A., Raicher, I., Gilberto A. Paula (2024) Additive partial linear models with autoregressive symmetric errors and its application to the hospitalizations for respiratory diseases Stat Papers 65, 5145–5166. doi:10.1007/s00362-024-01590-w

Oliveira, R.A., Paula, G.A. (2021) Additive partial linear models with autoregressive symmetric errors and its application to the hospitalizations for respiratory diseases Comput Stat 36, 2435–2466. doi:10.1007/s00180-021-01106-2

Examples

data(temperature)
datos = data.frame(temperature,time=1:length(temperature))
mod1<-aplms::aplms(temperature ~ 1,
                   npc=c("time"), basis=c("cr"),Knot=c(60),
                   data=datos,family=Powerexp(k=0.3),p=1,
                   control = list(tol = 0.001,
                                  algorithm1 = c("P-GAM"),
                                  algorithm2 = c("BFGS"),
                                  Maxiter1 = 20,
                                  Maxiter2 = 25),
                   lam=c(10))
#> [1] "Iteration 1"
#> Iteration 1 - 1[1] 1
#> Iteration 1 - 2[1] 2
#> Iteration 1 - 3[1] 3
#> Iteration 1 - 4[1] 4
#> Iteration 1 - 5[1] 5
#> Iteration 1 - 6[1] 6
#> Iteration 1 - 7[1] 7
#> [1] "Iteration 2"
#> Iteration 2 - 1[1] 1
#> Iteration 2 - 2[1] 2
#> Iteration 2 - 3[1] 3
#> Iteration 2 - 4[1] 4
#> Iteration 2 - 5[1] 5
#> Iteration 2 - 6[1] 6
#> [1] "Iteration 3"
#> Iteration 3 - 1[1] 1
#> Iteration 3 - 2[1] 2
#> Iteration 3 - 3[1] 3
#> [1] "Iteration 4"
#> Iteration 4 - 1[1] 1
summary(mod1)
#>  ---------------------------------------------------------------
#>  Additive partial linear models with symmetric errors 
#>  ---------------------------------------------------------------
#>  Sample size:  142 
#>  -------------------------- Model ---------------------------
#> 
#> aplms::aplms(formula = temperature ~ 1, npc = c("time"), basis = c("cr"), 
#>     Knot = c(60), data = datos, family = Powerexp(k = 0.3), p = 1, 
#>     control = list(tol = 0.001, algorithm1 = c("P-GAM"), algorithm2 = c("BFGS"), 
#>         Maxiter1 = 20, Maxiter2 = 25), lam = c(10))
#>  
#>  ------------------- Parametric component -------------------
#> 
#>           Estimate Std. Error t value  Pr(>|t|)    
#> intercept 0.056619     0.0041 13.8905 < 2.2e-16 ***
#> 
#>  ----------------- Non-parametric component ------------------ 
#> 
#>          Wald       df    Pr(>.)    
#> time 7589.838   58.583 < 2.2e-16 ***
#> 
#>  --------------- Autoregressive and Scale parameter ---------------- 
#> 
#>        Estimate Std. Error    Wald  Pr(>|t|)    
#> phi   0.0022992     0.0003  7.3902 1.242e-11 ***
#> rho1 -0.2571215     0.0662 -3.8866 0.0001568 ***
#>  
#> 
#> ------ Penalized Log-likelihood and Information criterion------
#> 
#>  Log-lik:  200.07 
#>  AIC    :  -276.97 
#>  AICc   :  -274.46 
#>  BIC    :  -94.94 
#>  GCV    :  0.01 
#> 
#>  --------------------------------------------------------------------
print(mod1)
#> Call:
#> aplms::aplms(formula = temperature ~ 1, npc = c("time"), basis = c("cr"), 
#>     Knot = c(60), data = datos, family = Powerexp(k = 0.3), p = 1, 
#>     control = list(tol = 0.001, algorithm1 = c("P-GAM"), algorithm2 = c("BFGS"), 
#>         Maxiter1 = 20, Maxiter2 = 25), lam = c(10))