Family Objects for Elliptical Models

The family object provide an specify details of the model APLMS. The distribution functions are necessary to specify the random component of the regression models with elliptical errors. The code is derived from the archived package gwer (Araujo, Y.A., Cysneiros, F.J.A., and Cysneiros, A.H.M.A., 2022), originally available on CRAN.

Usage

# S3 method for class 'elliptical'
family(object, ...)

Normal()

Cauchy()

LogisI()

LogisII()

Student(df = stop("no df argument"))

Powerexp(k = stop("no k argument"))

Glogis(parma = stop("no alpha=alpha(m) or m argument"))

Gstudent(parm = stop("no s or r argument"))

Cnormal(parmt = stop("no epsi or sigma argument"))

GNormal(nu = stop("no nu argument"))

Arguments

object

an object with the result of the fitted elliptical regression model.

...

arguments to be used to form the default control argument if it is not supplied directly.

df

degrees of freedom.

k

shape parameter.

parma

parameter vector (alpha, m).

parm

parameter vector (s, r) for this distribuition.

parmt

parameters vector (epsi, sigma).

nu

degrees of freedom.

Value

An object of class “family” specifying a list with the follows elements:

family

character: the family name.

g0, g1, g2, g3, g4, g5

derived fuctions associated with the distribution family defined.

df

degree of freedom for t-Student distribution.

s, r

shape parameters for generalized t-Student distribution.

alpha

shape parameter for contaminated normal and generalized logistic distributions.

mp

shape parameter for generalized logistic distribution.

epsi,sigmap

dispersion parameters for contaminated normal distribution.

k

shape parameter for power exponential distribution.

References

Fang, K. T., Kotz, S. and NG, K. W. (1990, ISBN:9781315897943). Symmetric Multivariate and Related Distributions. London: Chapman and Hall.

Examples

Normal()
#> 
#>  Normal family
#> 
#>  density  :    log(1/(sqrt(2 * pi)) * exp(-0.5 * z^2)) 
#> 
#>  Wg:    rep(-0.5, length(z)) 
#> 
#>  scale:    1/4 
#> 
#>  scale dispersion:    3/4 
#> 
#>  scale variance:    1 
#> 
#>  Wg':    rep(0, length(z)) 
Powerexp(k=0.1)
#> 
#>  Powerexp family
#> 
#>  density  :     log(1/(gamma(1 + ((1 + k)/2)) * 2^(1 + (1 + k)/2)) * exp(-0.5 * (abs(z)^(2/(1 + k))))) 
#> 
#>  Wg:     1/(-2 * (1 + k) * (z^2)^(k/(1 + k))) 
#> 
#>  scale:     (gamma((3 - k)/2))/(4 * (2^(k - 1) * (1 + k)^2 * gamma((k + 1)/2))) 
#> 
#>  scale dispersion:     (k + 3)/(4 * (k + 1)) 
#> 
#>  scale variance:     2^(1 + k) * (gamma(1.5 * (k + 1))/(gamma((k + 1)/2))) 
#> 
#>  Wg':     k/(2 * (z^2)^((2 * k + 1)/(1 + k)) * ((1 + k)^2)) 
#> 
#>  k  : 0.1