Respiratory diseases hospitalization Dataset

This dataset consists of respiratory diseases hospitalization in Sorocaba, São Paulo, Brazil. The details of the statistical modeling using the APLMS-AR(p) approach can be found in Chou-Chen, et al. (2024). doi:10.1007/s00362-024-01590-w . The hospitalization data of respiratory diseases in Sorocaba city, São Paulo, Brazil are obtained from the Hospital Information System of Brazil’s Unified National Health System (SIH-SUS), and the climatic and pollution data are provided by the QUALAR system.

Usage

data(hospitalization)

Format

The "data" slot is a data frame with 932 weekly data on the following 29 variables.

date, year, epi.week, tdate

Date, year, epidemiologic weeks, and time index.

y

Respiratory hospitalization count.

MP10_max, MP10_min, MP10_avg

Maximum, minimum and average of \(MP10\).

NO_max,NO_min,NO_avg

Maximum, minimum and average of \(NO\).

NO2_max, NO2_min, NO2_avg

Maximum, minimum and average of \(NO_2\).

NOx_max, NOx_min, NOx_avg

Maximum, minimum and average of \(NO_x\).

O3_max, O3_min, O3_avg

Maximum, minimum and average of \(O_3\).

TEMP_max, TEMP_min, TEMP_avg

Maximum, minimum and average of temperature.

RH_max, RH_min, RH_avg

Maximum, minimum and average of relative humidity.

ampl_max, ampl_min, ampl_avg

Maximum, minimum and average of daily temperature amplitude.

References

Chou-Chen, S.W., Oliveira, R.A., Raicher, I. et al. (2024) Additive partial linear models with autoregressive symmetric errors and its application to the hospitalizations for respiratory diseases. Stat Papers 65, 5145–5166. doi:10.1007/s00362-024-01590-w

Examples

data(hospitalization)
head(hospitalization)
#>         date year epi.week tdate  y MP10_max MP10_min MP10_avg NO_max NO_min
#> 1 2005-01-03 2005        1     1 64 167.9242 11.65520 44.51816     28      0
#> 2 2005-01-10 2005        2     2 58 159.9140 12.58550 46.16039     20      0
#> 3 2005-01-17 2005        3     3 62 158.2932 12.04895 45.28987     61      0
#> 4 2005-01-24 2005        4     4 53 169.7507 12.23643 45.59006     56      0
#> 5 2005-01-31 2005        5     5 52 181.2455 13.03868 48.15511     39      0
#> 6 2005-02-07 2005        6     6 47 161.9691 12.68160 47.74442     37      0
#>     NO_avg NO2_max NO2_min   NO2_avg NOx_max NOx_min   NOx_avg O3_max O3_min
#> 1 4.886598      37       5 15.649485      42       3 12.360825     77      1
#> 2 2.664596      41       4 14.857143      38       2 10.105590    124      1
#> 3 4.956522      36       3 13.000000      62       2 11.012422    114      1
#> 4 6.579710      48       2 15.413043      51       1 13.550725     78      0
#> 5 2.732919      24       1  8.118012      44       1  6.596273    119      2
#> 6 3.496894      72       3 13.472050      54       2 10.099379    157      2
#>     O3_avg TEMP_max TEMP_min TEMP_avg RH_max RH_min   RH_avg ampl_max ampl_min
#> 1 25.25000     30.7     19.1 23.25119     99     55 92.04167     11.3      4.9
#> 2 41.54375     32.2     17.1 24.47844     99     52 84.89881     13.3      4.8
#> 3 26.22981     29.5     20.1 23.06071     99     70 95.29762      8.2      2.3
#> 4 26.74534     30.8     17.8 22.13929     99     67 92.19643     11.3      3.3
#> 5 25.32298     30.3     18.1 22.86190     98     51 85.04790     10.9      6.0
#> 6 47.27950     29.6     16.7 21.99821     99     44 74.63095     11.9      9.0
#>    ampl_avg
#> 1  7.800000
#> 2  8.857143
#> 3  5.600000
#> 4  6.314286
#> 5  7.400000
#> 6 10.557143